requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
databases[aiosqlite]>=0.8.0
aiosqlite>=0.19.0
libsql-client>=0.5.0
pydantic>=2.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0